+++ test/chat_models/chat_bumblebee_test.exs
@@ -5,6 +5,7 @@ defmodule LangChain.ChatModels.ChatBumblebeeTest do
   doctest LangChain.ChatModels.ChatBumblebee
   alias LangChain.ChatModels.ChatBumblebee
   alias LangChain.Message
+  alias LangChain.TokenUsage
 
   describe "new/1" do
     test "works with minimal attr" do
@@ -23,7 +24,7 @@ defmodule LangChain.ChatModels.ChatBumblebeeTest do
 
   describe "call/4" do
     test "supports API override" do
-      set_api_override({:ok, [Message.new_assistant!("Colorful Threads")]})
+      set_api_override({:ok, [Message.new_assistant!("Colorful Threads")], :on_llm_new_message})
 
       # https://js.langchain.com/docs/modules/models/chat/
       {:ok, chat} = ChatBumblebee.new(%{serving: Fake})
@@ -39,15 +40,23 @@ defmodule LangChain.ChatModels.ChatBumblebeeTest do
 
   describe "do_process_response/3" do
     setup do
-      callback_fn = fn data ->
-        send(self(), {:callback_data, data})
-      end
+      handler = %{
+        on_llm_new_delta: fn _model, delta ->
+          send(self(), {:callback_delta, delta})
+        end,
+        on_llm_new_message: fn _model, message ->
+          send(self(), {:callback_message, message})
+        end,
+        on_llm_token_usage: fn _model, usage ->
+          send(self(), {:callback_usage, usage})
+        end,
+      }
 
-      %{callback_fn: callback_fn}
+      %{handler: handler}
     end
 
-    test "handles non-streamed full text response", %{callback_fn: callback_fn} do
-      model = ChatBumblebee.new!(%{serving: Fake, stream: false})
+    test "handles non-streamed full text response", %{handler: handler} do
+      model = ChatBumblebee.new!(%{serving: Fake, stream: false, callbacks: [handler]})
 
       response = %{
         results: [
@@ -60,33 +69,33 @@ defmodule LangChain.ChatModels.ChatBumblebeeTest do
 
       expected_message = Message.new_assistant!(%{content: "Hello.", status: :complete})
 
-      [message] = ChatBumblebee.do_process_response(response, model, callback_fn)
+      [message] = ChatBumblebee.do_process_response(response, model)
       assert message == expected_message
 
-      assert_received {:callback_data, data}
+      assert_received {:callback_message, data}
       assert data == expected_message
     end
 
-    test "handles stream when stream: false", %{callback_fn: callback_fn} do
-      model = ChatBumblebee.new!(%{serving: Fake, stream: false})
+    test "handles stream when stream: false", %{handler: handler} do
+      model = ChatBumblebee.new!(%{serving: Fake, stream: false, callbacks: [handler]})
 
       expected_message = Message.new_assistant!(%{content: "Hello.", status: :complete})
 
       stream =
-        ["He", "ll", "o", ".", {:done, %{input: 38, output: 4, padding: 4058}}]
+        ["He", "ll", "o", ".", {:done, %{token_summary: %{input: 38, output: 4, padding: 4058}}}]
         |> Stream.map(& &1)
 
-      [message] = ChatBumblebee.do_process_response(stream, model, callback_fn)
+      [message] = ChatBumblebee.do_process_response(stream, model)
       assert message == expected_message
 
-      assert_received {:callback_data, data}
+      assert_received {:callback_message, data}
       assert data == expected_message
     end
 
     test "handles a stream when stream: false and no stream_done requested", %{
-      callback_fn: callback_fn
+      handler: handler
     } do
-      model = ChatBumblebee.new!(%{serving: Fake, stream: false})
+      model = ChatBumblebee.new!(%{serving: Fake, stream: false, callbacks: [handler]})
 
       expected_message = Message.new_assistant!(%{content: "Hello.", status: :complete})
 
@@ -94,17 +103,17 @@ defmodule LangChain.ChatModels.ChatBumblebeeTest do
         ["He", "ll", "o", "."]
         |> Stream.map(& &1)
 
-      [message] = ChatBumblebee.do_process_response(stream, model, callback_fn)
+      [message] = ChatBumblebee.do_process_response(stream, model)
       assert message == expected_message
 
-      assert_received {:callback_data, data}
+      assert_received {:callback_message, data}
       assert data == expected_message
     end
 
     test "handles a stream when stream: true and no stream_done requested", %{
-      callback_fn: callback_fn
+      handler: handler
     } do
-      model = ChatBumblebee.new!(%{serving: Fake, stream: true})
+      model = ChatBumblebee.new!(%{serving: Fake, stream: true, callbacks: [handler]})
 
       expected_deltas = [
         MessageDelta.new!(%{content: "Hel", status: :incomplete, role: :assistant}),
@@ -115,20 +124,20 @@ defmodule LangChain.ChatModels.ChatBumblebeeTest do
         ["Hel", "lo."]
         |> Stream.map(& &1)
 
-      assert [deltas] = ChatBumblebee.do_process_response(stream, model, callback_fn)
+      assert [deltas] = ChatBumblebee.do_process_response(stream, model)
       assert deltas == expected_deltas
 
-      assert_received {:callback_data, data_1}
+      assert_received {:callback_delta, data_1}
       assert data_1 == MessageDelta.new!(%{content: "Hel", role: :assistant, status: :incomplete})
 
-      assert_received {:callback_data, data_2}
+      assert_received {:callback_delta, data_2}
       assert data_2 == MessageDelta.new!(%{content: "lo.", role: :assistant, status: :incomplete})
 
-      refute_received {:callback_data, _data_3}
+      refute_received {:callback_delta, _data_3}
     end
 
-    test "handles stream when stream: true", %{callback_fn: callback_fn} do
-      model = ChatBumblebee.new!(%{serving: Fake, stream: true})
+    test "handles stream when stream: true", %{handler: handler} do
+      model = ChatBumblebee.new!(%{serving: Fake, stream: true, callbacks: [handler]})
 
       expected_deltas = [
         %MessageDelta{content: "He", status: :incomplete, role: :assistant},
@@ -139,27 +148,30 @@ defmodule LangChain.ChatModels.ChatBumblebeeTest do
       ]
 
       stream =
-        ["He", "ll", "o", ".", {:done, %{input: 38, output: 4, padding: 4058}}]
+        ["He", "ll", "o", ".", {:done, %{token_summary: %{input: 38, output: 4, padding: 4058}}}]
         |> Stream.map(& &1)
 
       # expect the deltas to be in an outer
-      assert [deltas] = ChatBumblebee.do_process_response(stream, model, callback_fn)
+      assert [deltas] = ChatBumblebee.do_process_response(stream, model)
       assert deltas == expected_deltas
 
-      assert_received {:callback_data, data_1}
+      assert_received {:callback_delta, data_1}
       assert data_1 == MessageDelta.new!(%{content: "He", role: :assistant, status: :incomplete})
 
-      assert_received {:callback_data, data_2}
+      assert_received {:callback_delta, data_2}
       assert data_2 == MessageDelta.new!(%{content: "ll", role: :assistant, status: :incomplete})
 
-      assert_received {:callback_data, data_3}
+      assert_received {:callback_delta, data_3}
       assert data_3 == MessageDelta.new!(%{content: "o", role: :assistant, status: :incomplete})
 
-      assert_received {:callback_data, data_4}
+      assert_received {:callback_delta, data_4}
       assert data_4 == MessageDelta.new!(%{content: ".", role: :assistant, status: :incomplete})
 
-      assert_received {:callback_data, data_5}
+      assert_received {:callback_delta, data_5}
       assert data_5 == MessageDelta.new!(%{role: :assistant, status: :complete})
+
+      assert_received {:callback_usage, usage}
+      assert %TokenUsage{input: 38, output: 4} = usage
     end
   end
 end